+++ docs/embed.ts
@@ -1,5 +1,7 @@
 import sdk from '@stackblitz/sdk';
 
+// https://api.npmjs.org/versions/@hey-api%2Fopenapi-ts/last-week
+
 export const embedProject = (projectId: string) => async (event: Event) => {
   const container = document.createElement('div');
 
@@ -8,12 +10,20 @@ export const embedProject = (projectId: string) => async (event: Event) => {
     node.replaceWith(container);
   }
 
-  if (projectId === 'hey-api-example') {
-    return await sdk.embedProjectId(container, projectId, {
-      height: 700,
-      openFile:
-        'openapi-ts.config.ts,src/client/schemas.gen.ts,src/client/services.gen.ts,src/client/types.gen.ts',
-      view: 'editor',
-    });
+  switch (projectId) {
+    case 'hey-api-client-fetch-example':
+      return await sdk.embedProjectId(container, projectId, {
+        height: 700,
+        openFile:
+          'openapi-ts.config.ts,src/client/schemas.gen.ts,src/client/services.gen.ts,src/client/types.gen.ts,src/App.tsx',
+        view: 'editor',
+      });
+    case 'hey-api-example':
+      return await sdk.embedProjectId(container, projectId, {
+        height: 700,
+        openFile:
+          'openapi-ts.config.ts,src/client/schemas.gen.ts,src/client/services.gen.ts,src/client/types.gen.ts',
+        view: 'editor',
+      });
   }
 };