+++ packages/client-core/src/types.ts
@@ -7,24 +7,26 @@ export type ApiResult<TData = any> = {
 };
 
 export type ApiRequestOptions = {
-  readonly method:
-    | 'GET'
-    | 'PUT'
-    | 'POST'
-    | 'DELETE'
-    | 'OPTIONS'
-    | 'HEAD'
-    | 'PATCH';
-  readonly url: string;
-  readonly path?: Record<string, unknown>;
-  readonly cookies?: Record<string, unknown>;
-  readonly headers?: Record<string, unknown>;
-  readonly query?: Record<string, unknown>;
-  readonly formData?: Record<string, unknown>;
   readonly body?: any;
+  readonly cookies?: Record<string, unknown>;
+  readonly errors?: Record<number | string, string>;
+  readonly formData?: Record<string, unknown>;
+  readonly headers?: Record<string, unknown>;
   readonly mediaType?: string;
+  readonly method:
+    | 'CONNECT'
+    | 'DELETE'
+    | 'GET'
+    | 'HEAD'
+    | 'OPTIONS'
+    | 'PATCH'
+    | 'POST'
+    | 'PUT'
+    | 'TRACE';
+  readonly path?: Record<string, unknown>;
+  readonly query?: Record<string, unknown>;
   readonly responseHeader?: string;
-  readonly errors?: Record<number, string>;
+  readonly url: string;
 };
 
 export type Headers = Record<string, string>;